+++ src/index.ts
@@ -1,4 +1,4 @@
-import { renderInitialData, renderSearchForm } from './ts/renderData';
+import { renderInitialData, renderSearchForm } from './ts/App';
 
 renderInitialData();
 renderSearchForm();
